+++ src/main/java/com/sprint/sb06deokhugamteam01/dto/book/naver/NaverBookSearchResponse.java
@@ -1,0 +1,10 @@
+package com.sprint.sb06deokhugamteam01.dto.book.naver;
+
+public record NaverBookSearchResponse(
+        String lastBuildDate,
+        int total,
+        int start,
+        int display,
+        BookData[] items
+) {
+}
