+++ src/main/java/com/sprint/sb06deokhugamteam01/service/book/NaverBookSearchService.java
@@ -1,8 +1,9 @@
 package com.sprint.sb06deokhugamteam01.service.book;
 
+import com.sprint.sb06deokhugamteam01.dto.book.naver.BookData;
 import com.sprint.sb06deokhugamteam01.dto.book.BookDto;
+import com.sprint.sb06deokhugamteam01.dto.book.naver.NaverBookSearchResponse;
 import com.sprint.sb06deokhugamteam01.exception.book.BookInfoFetchFailedException;
-import com.sprint.sb06deokhugamteam01.exception.book.InvalidIsbnException;
 import com.sprint.sb06deokhugamteam01.exception.book.NoSuchBookException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -11,7 +12,6 @@ import org.springframework.http.HttpStatusCode;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClient;
 
-import java.time.LocalDate;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -45,52 +45,21 @@ public class NaverBookSearchService implements BookSearchService{
                 }))
                 .body(NaverBookSearchResponse.class);
 
-        if (result.items.length == 0) {
+        if (result.items().length == 0) {
             throw new NoSuchBookException(detailMap("isbn", isbn));
         }
 
         BookData bookData = result.items()[0];
 
         return BookDto.builder()
-                .title(bookData.title)
-                .author(bookData.author)
-                .description(bookData.description)
-                .publisher(bookData.publisher)
+                .title(bookData.title())
+                .author(bookData.author())
+                .description(bookData.description())
+                .publisher(bookData.publisher())
                 .publishedDate(bookData.getPublishedDate())
-                .isbn(bookData.isbn)
-                .thumbnailUrl(bookData.image)
+                .isbn(bookData.isbn())
+                .thumbnailUrl(bookData.image())
                 .build();
-    }
-
-    private record NaverBookSearchResponse(
-            String lastBuildDate,
-            int total,
-            int start,
-            int display,
-            BookData[] items
-    ) {}
-
-    private record BookData(
-            String title,
-            String link,
-            String image,
-            String author,
-            String discount,
-            String publisher,
-            String pubdate,
-            String isbn,
-            String description
-    ) {
-        public LocalDate getPublishedDate() {
-            if (pubdate.length() != 8) {
-                return null;
-            }
-            return LocalDate.of(
-                    Integer.parseInt(pubdate.substring(0, 4)),  // year
-                    Integer.parseInt(pubdate.substring(4, 6)),  // month
-                    Integer.parseInt(pubdate.substring(6, 8))   // day
-            );
-        }
     }
 
     private Map<String, Object> detailMap(String key, Object value) {
